🔧 Move modelConfig.tsx to upper folder 🐛 Optimize the loading timing of the model status
